Keep benchmark tooling honest and remove perf-runner duplication
The perf follow-up had drifted into overstated naming and split logic between two near-identical runners. This change renames the loop to match what it actually does, moves the benchmark orchestration under benchmarks/, removes the misleading autoresearch doc, and extracts the shared baseline/verification/run helpers into one perf library.

Constraint: The current loop is a benchmark iteration harness, not an autonomous research system
Constraint: Keep the cleanup scoped to the benchmark/perf runner surface while preserving the JIT-friendly runtime work already landed
Rejected: Keep the autoresearch naming and just soften docs | still leaves the command/file layout misleading
Rejected: Leave perf:bench and perf:loop duplicated in separate files | needless maintenance slop on a new surface
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep benchmark orchestration colocated with benchmarks unless it grows beyond measurement-only responsibilities
Tested: deno fmt on benchmark runner files; deno check benchmarks/_perf_lib.ts benchmarks/perf-bench.ts benchmarks/perf-loop.ts; deno task perf:bench -- --skip-verify --runtime preview-host --project perfcase; deno task perf:loop -- --runs 1 --skip-verify --runtime preview-host --project perfcase
Not-tested: full repo typecheck/test suite after this cleanup-only rename/move pass
