+++ deno.json
@@ -358,8 +358,8 @@
     "bench:browser": "deno run -A benchmarks/browser/run.ts",
     "bench:server": "deno run -A benchmarks/server/run.ts",
     "bench:compare:local": "deno run -A benchmarks/compare/local.ts",
-    "perf:bench": "deno run -A auto/bench.ts",
-    "perf:autoresearch": "deno run -A auto/autoresearch.ts",
+    "perf:bench": "deno run -A benchmarks/perf-bench.ts",
+    "perf:loop": "deno run -A benchmarks/perf-loop.ts",
     "check:circular": "deno run -A jsr:@cunarist/deno-circular-deps src/index.ts",
     "cli": "deno run --allow-all cli/main.ts",
     "mcp": "deno run --allow-all cli/main.ts mcp",
